refactor: record/excption 오타 정정
