+++ src/main/java/com/umc/halo/domain/record/service/RecordService.java
@@ -16,8 +16,8 @@ import com.umc.halo.domain.record.converter.*;
 import com.umc.halo.domain.record.dto.*;
 import com.umc.halo.domain.record.entity.*;
 import com.umc.halo.domain.record.enums.*;
-import com.umc.halo.domain.record.excption.*;
-import com.umc.halo.domain.record.excption.code.*;
+import com.umc.halo.domain.record.exception.*;
+import com.umc.halo.domain.record.exception.code.*;
 import com.umc.halo.domain.record.repository.*;
 import com.umc.halo.global.ai.event.*;
 import lombok.*;
